requirements-colab.txt
torch
esm
biopython
tqdm
requests
